Django Rest Basics/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-health-check==3.16.5
djangorestframework==3.13.1
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.0.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7